+++ assets/statistics.js
@@ -1,0 +1,44 @@
+$(document).on("rex:ready", function (event, container) {
+    statistics_datefilter_start = document.getElementById("statistics_datefilter_start");
+    statistics_df_lsd = document.getElementById("statistics_df_lsd");
+    statistics_df_ltd = document.getElementById("statistics_df_ltd");
+    statistics_df_ty = document.getElementById("statistics_df_ty");
+    statistics_df_wt = document.getElementById("statistics_df_wt");
+
+    last_seven_days = get_past_date(7);
+    last_thirty_days = get_past_date(30);
+    this_year = get_this_year();
+    whole_time = statistics_df_wt.getAttribute("data-start");
+
+    statistics_df_lsd.addEventListener("click", function () {
+        statistics_datefilter_start.value = last_seven_days;
+    });
+    statistics_df_ltd.addEventListener("click", function () {
+        statistics_datefilter_start.value = last_thirty_days;
+    });
+    statistics_df_ty.addEventListener("click", function () {
+        statistics_datefilter_start.value = this_year;
+    });
+    statistics_df_wt.addEventListener("click", function () {
+        statistics_datefilter_start.value = whole_time;
+    });
+
+    function get_past_date(minusDays) {
+        var date = new Date();
+        date.setDate(date.getDate() - minusDays);
+        day = ("0" + date.getDate()).slice(-2);
+        month = ("0" + date.getMonth()).slice(-2);
+        year = date.getFullYear();
+        str = year + "-" + month + "-" + day;
+
+        return str;
+    }
+
+    function get_this_year() {
+        var date = new Date();
+        year = date.getFullYear();
+        str = year + "-01-01";
+
+        return str;
+    }
+});